groups and an environment to support the cloud product
This adds two groups and an environment.

The "cloud-server" group contains packages required to support running
Fedora inside of cloud environments.  It is meant to be used for building
cloud images.

The "cloud-server-environment" environment pulls in the above groups as
well as core.  It is meant to be used to enable an interactive install
of a cloud-like image for development and testing.

The "cloud-management" group contains packages that enable management
of images and VMs running inside of clouds.