fedora-repos-anaconda needs to be pulled in via lorax, not spin-kickstarts
also adjust the comment on @anaconda-tools (which handily explains exactly
this issue, but which we clearly forgot to read...) to mention lorax as
another way stuff gets pulled into the anaconda runtime environment. We
really need to clear up what should be a dep of anaconda and what should be
included in lorax.