add a 'ansible-node' group
As per recent discussion on fedora-devel and in Server WG
meeting today, this seems like a good idea: a package group
whose purpose is to ensure the system can be managed by ansible
without special measures. Right now it just contains python2-
dnf as that's what you need in place to be able to install
other packages via ansible.