Move non-core apps from gnome-desktop to workstation-product
These apps are not part of the upstream GNOME desktop; they are where we
diverge from upstream, so they belong in the workstation-product group.
Of these apps, Bijiben is desired to eventually become a core GNOME app,
but it's not very good yet and it's recommended that distributions not
ship it. Most users will be better off with webmail than with Evolution.
Rhythmbox will eventually be replaced by GNOME Music, and Shotwell by
GNOME Photos. Boxes has too many bugs right now, but it's also intended
for core eventually.