Move seahorse to workstation-product
It's not in GNOME core, so it's a bug that it's in the gnome-desktop
group. Keeping it installed is deviation from upstream, so it belongs in
workstation-product. Make the change only for F26 since there's no
immediate benefit and spins might be surprised for it to disappear this
late in the game.

We hope to move it to GNOME core in the future, after a rewrite.