add sort target to Makefile and sort
As people are always getting things out of order in comps. add a make
target that sorts all the comps files, this way people can just run
"make sort" when making changes, and we can add a jenkins job that
will fail and mark a PR as bad for unsorted commits. As the comps
files are really badly sorted now there is some massive change for
EOL comps fils in here. but the run is needed so we can pass going
forward

Signed-off-by: Dennis Gilmore <dennis@ausil.us>