Include sscg by default alongside cockpit
Cockpit 140+ will now use sscg if it is available to generate more
reliable and secure default certificates instead of simple
self-signed certificates.

Fedora Server carries only a reduced set of Cockpit subpackages,
so it doesn't pick up the `Recommends: sscg` from the main
'cockpit' metapackage. Thus we need to add it explicitly in the
comps.xml

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>