Tweak the updated grub2 entries into alphabetical order
I forgot about this, but there's something which likes comps
groups to be in perfect alphabetical order...@kevin ran into
some mysterious change on the compose boxes which corrected the
order of the grub2 entries after some of the changes we made
recently, which is probably to do with that. So let's fix the
recently-changed grub lines to be in proper alphabetical order.