Make core Workstation packages mandatory
This makes core Workstation components, including the 'Core
Applications' from the technical specification, mandatory in
comps. This means that if they are removed, the group they are
in (gnome-desktop or firefox) is no longer considered to be
installed, and compose of any images containing these groups
will fail if any of the packages cannot be installed (which is
the main goal of the change; to avoid the images composing but
missing these key packages).

Signed-off-by: Adam Williamson <awilliam@redhat.com>