Make core Xfce components mandatory
This makes core Xfce desktop packages mandatory. That means
the group they are in (xfce-desktop) will not be considered to be
'installed' if they are removed, and importantly, image composes
will fail (as opposed to 'succeeding', but with the packages left
out) if they are missing or not installable. That's the main
benefit here.

Amendments to the list welcome, this was just my best guess at
what constitutes the really core packages.

Signed-off-by: Adam Williamson <awilliam@redhat.com>