Make core libreoffice packages mandatory
This makes core LibreOffice packages mandatory. That means the
group they are in (libreoffice) will not be considered to be
'installed' if they are removed, and importantly, image composes
will fail (as opposed to 'succeeding', but with the packages left
out) if they are missing or not installable. That's the main
benefit here.

Signed-off-by: Adam Williamson <awilliam@redhat.com>