Drop firmware-addon-dell from comps (#1561768)
firmware-addon-dell is currently uninstallable and FTBFS in F28
and Rawhide. I also suspect that it may be a dead and useless
package - see https://bugzilla.redhat.com/show_bug.cgi?id=1555752
(the FTBFS bug).

Until that can all be resolved, at least, let's drop it from
comps. It is currently getting pulled into the Server DVD, which
is technically a release-blocking issue as long as it is not
installable.

Signed-off-by: Adam Williamson <awilliam@redhat.com>