Use cockpit metapackage, add cockpit-packagekit
The cockpit metapackage is meant to control the minimal and recommended
parts of the default installation. This replaces
cockpit-{bridge,shell,ws}.

Also add cockpit-packagekit, as that is the enabler for installing other
cockpit extensions and required packages that Cockpit can control (such
as nfs-utils, firewalld, dnf-automatic, or pcp), and provides Software
Updates, which is an important system administration task on every kind
of installation.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1628457