Add a little script to check for missing packages in comps
It would be nice to make 'package is listed in group, but does
not exist' a fatal error (at least optionally) for dnf etc, but
we can't really do that at present because it's *not* been a
fatal error for so long that our comps have tons of stale
entries for packages that no longer exist, and also entries that
aren't properly archified for packages which only exist on some
arches. This script pokes through the comps file for current
Rawhide and identifies all (I hope) pkgreqs which specify a
package that is not actually available in current Rawhide, per
arch.

Signed-off-by: Adam Williamson <awilliam@redhat.com>