Move grubby and grubby-deprecated from @core to @arm-tools
Adding them into @core was supposed to make them show up in some
install trees, according to the commit message, but it broke
regular BLS installs:

https://bugzilla.redhat.com/show_bug.cgi?id=1654036

which is obviously a major problem. I *think* moving the entries
to arm-tools should ensure the packages are in the relevant trees
without breaking BLS installs.

Signed-off-by: Adam Williamson <awilliam@redhat.com>