Move grubby from @core to @standard
With BLS the grubby tool isn't used when new kernels are installed, since
the bootloader config files don't have to be edited anymore. But u-boot
doesn't have BLS support yet, so on ARMv7 installs the old grubby tool
needs to be installed.

That's why commit 4c2694e1d8b ("Move grubby and grubby-deprecated from
@core to @arm-tools") moved the grubby-depracted package to @arm-tools.

But it also moved the grubby package and while this isn't needed during
install on other arches, it may be useful for users to manage their boot
menu entries.

Suggested-by: Adam Williamson <awilliam@redhat.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>