Remove grubby-deprecated from @arm-tools
Most bootloaders used in Fedora already have BootLoaderSpec, but extlinux
doesn't yet. So when using extlinux the old grubby must to be installed.

The grubby-deprecated package was added to the @arm-tools group to ensure
that it will be installed on ARMv7 machines. But on aarch64 machines this
group can also be installed, which means that grubby-deprecated will be
leading to Anaconda disabling the BLS support.

Instead, the right place to solve this is in Anaconda which already does
install grubby-deprecated when the extlinux bootloader option is chosen.

That way grubby-deprecated will only be installed for ARMv7 machines, but
won't be installed in aarch64 machines that use grub2 EFI and support BLS.

Resolves: rhbz#1678843

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>