@buildsys-build needs to rely on fedora-release-common
Attempting to install @buildsys-build atop a Workstation, Server or
other edition/spin using a fedora-release-$VARIANT package resulted
in a conflict because @buildsys-build unconditionally required
fedora-release. We will change it to fedora-release-common which
is guaranteed to be present (and will pull in fedora-release, in
the case of a clean buildroot).

Resolves: rhbz#1728863

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>