Remove epiphany group
We don't need two web browsers in the gnome-desktop-environment group. I
think I have a pretty credible claim to be Epiphany's number 1 fan, but
Firefox is our default. Users who want Epiphany can install it manually.

This means we no longer have any need for the Epiphany comps group.
There's no need to have a comps group for exactly one package if it's
not going to be included in any category.