Move back grubby to @core group
This package used to be in the @core group since it was needed for kernel
installs, but was moved to the @standard group since isn't needed anymore.

The package is still needed though to install kernels from source, since
it ships the installkernel script that is invoked by the kernel Makefile.

It also has a grubby script that allows to choose a default kernel, change
the kernel command line parameters and manage the boot entries in general.

But the @standard group is not included in all the products and spins. For
example is not included in the Fedora Workstation Product.

So move the grubby package again to the @core group and make it mandatory,
since still provides essential functionality and should always be present.

Reported-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>