Make libreoffice-math optional
LibreOffice Math is not an approved default app in Fedora Workstation.
We shipped it in F32 anyway by mistake. The .desktop file was changed to
no longer be hidden, and the -math subpackage was changed to include only
the desktop file and not the actual math application so that it wouldn't
appear as an installed application unless the subpackage were to be
installed, but it's being installed by default so the plan was defeated.
See rbhz#1776774 for some history.

Anyway, solution is to stop installing this package. LibreOffice Math
will still actually be installed, it'll just *look* like it's not. The
goal is to display only Writer, Calc, and Impress, as before.