Drop plymouth from @core and @standard (#1933378)
As discussed in the bug report, the fallback text theme for
Plymouth is under-maintained and buggy. Plymouth maintainers
would prefer we use the graphical theme or not use Plymouth at
all. Consensus seems to support using graphical theme for
graphical installs and no Plymouth at all for non-graphical
installs, so this aims to achieve that, by dropping plymouth
from @core and @standard but keeping plymouth-system-theme in
@base-x and changing the plymouth in @workstation-product to
plymouth-system-theme (this isn't strictly necessary, but makes
things clear and consistent).

Signed-off-by: Adam Williamson <awilliam@redhat.com>