Put Plymouth back in @core (as default)
This is a partial revert of 61f4c2a, which removed plymouth from
@core and @standard. That was redundant, so this just puts it
back in @core, and now as default not mandatory.

The main justification for putting it back is
https://bugzilla.redhat.com/show_bug.cgi?id=1940163 - without
Plymouth installed, console output during boot on default aarch64
deployments is not shown on a local display even if one is
connected. This is particularly unfortunate in the case of
passphrase prompts for encrypted storage devices.

We believe the bug that caused us to take it out in the first
place - https://bugzilla.redhat.com/show_bug.cgi?id=1933378 -
should now be resolved, so putting it back should not cause that
to break again. If it does, we can always revert this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>