Move systemd-oomd-defaults into specific environment groups
It is removed from @core, and added to @cloud-server, @gnome-desktop,
@kde-desktop, @server-product, @workstation-product instead.

My understanding is that both KDE and GNOME use cgroups and work
nicely with systemd-oomd, and cloud and server should be fine with it
too. The environments that use those groups previously included
systemd-oomd-defaults from @core, so this change should not add
systemd-oomd-defaults to anything, but remove it from the groups and
environments that don't include the five groups that now reference it.

Fixes https://pagure.io/fedora-comps/issue/737.

Not tested.