Update KDE critical path definition for F37 and Rawhide
I've had a ticket for this for months:
https://pagure.io/fedora-kde/SIG/issue/171
but it really needs doing so we are testing and gating critical
KDE updates, so I just went ahead and came up with something.
This is based on the recent Plasma 5.24.5 update for Fedora 35,
I just went through the packages in that and picked out the
ones that looked most critical.

Signed-off-by: Adam Williamson <awilliam@redhat.com>