check-missing: don't print optional packages with missing arches
It doesn't make sense to me to worry about specifying the
correct arch list for optional packages. Optional packages are
optional by definition, so it seems fine to just specify them
and if they don't exist on the arch right now, oh well. This
way, if they do start building for the arch in future, they will
be picked up right away with no need to change comps.

Signed-off-by: Adam Williamson <awilliam@redhat.com>