Drop explicit listing of libcanberra-gtk3
This seems to be a hangover from years ago when it somehow had
to be explicitly listed to ensure the right sound theme got used
or something. These days many mandatory bits of GNOME, including
gnome-settings-daemon, require it, so it really doesn't need to
be specifically listed in its own right.

Signed-off-by: Adam Williamson <awilliam@redhat.com>