Split critical-path-base
This is intended to help with openQA test focus. We split up the
base group by purpose: things for installing (anaconda), things
for building packages (build), things for building deliverables
(compose), and things for runtime (base). This should allow us
to avoid unnecessarily running the openQA tests on e.g. gcc
updates (which is a useless waste of time), and potentially allow
us to run different/extended tests on e.g. anaconda updates. I
intentionally didn't add additional packages to the split groups
yet, we can do that as a follow-up, but I did replace livecd-tools
with lorax in the compose group, as we use things from lorax for
creating images now.

Signed-off-by: Adam Williamson <awilliam@redhat.com>