Add cups-browsed into ELN and F38 comps
There will be a new package - cups-browsed - which was previous part of
cups-filters package. It is useful for automatic installation of many
remote printers (usual use case on server), so it would be great if the
package was added to `print-server` (in ELN) and `printing` (in F38)
comps groups.

This is a part of self-contained change
https://bugzilla.redhat.com/show_bug.cgi?id=2166687