Remove grubby from core, enable systemd-boot installs
With the advent of alternative bootloaders (systemd-boot) whether
or not grubby gets installed should be dependent on whether grub
is selected as the bootloader. This currently happens correctly
with anaconda and the grub dependencies, so it can be removed
safely. That allows systemd-boot to be installed cleanly without
grubby drippings confusing it.

While anaconda may now request sdubby and systemd-boot, i'm
told the comment in the anaconda section isn't 100% accurate
as anaconda may try to install all those selections to the live
images. So, lets not add those to the anaconda comps yet.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>