Drop gnome-themes-extra from GNOME groups
It's an old package that used to ship empty css files to let
gnome-tweaks know which gtk3 themes are available (after gtk3 dropped
the standalone css files and started building them in).

This is now no longer needed since the high contrast and dark themes are
available from gnome-control-center, which does not use the css file
craziness to detect what themes are available.

It also used to pull in the gtk2 adwaita theme, but that's now handled
by conditional recommends from the gtk2 package instead.

Signed-off-by: Kalev Lember <klember@redhat.com>