CI: explain better what "git status failed!" means
`make sort` won't usually *fail*, even if it sorts stuff (it'll
only fail if it actually can't sort for some reason). So when
either the PR or the existing state of the repo isn't properly
sorted, what we get is the "git status failed!" case. I *think*
this is likely the only time we'll get that, I don't think
"make validate" would actually change files without failing. So
let's gloss that a git status failure probably means a sorting
issue.

Signed-off-by: Adam Williamson <awilliam@redhat.com>