F40:  Remove grubby from core, enable systemd-boot installs
With the advent of alternative bootloaders (systemd-boot) whether
or not grubby gets installed should be dependent on whether grub
is selected as the bootloader. This currently happens correctly
with anaconda and the grub dependencies, so it can be removed
safely. That allows systemd-boot to be installed cleanly without
grubby drippings confusing it.

Sync F40 with F39 since the branch happened before the commit was
merged.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>