f40: Add sdubby to list of package that may be needed by anaconda
Now that the sdubby has been reviewed, it needs to be included
in the list of packages that anaconda _may_ install on the target
machine. It should not be installed in the install media/etc
which is grub based, and is explicitly marked as conflicting with
grubby to avoid that.

Futher to avoid grubby and sdubby from both being installed on live
media there is a change in fedora-live-base.ko which excludes
sdubby from live media.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>