anaconda-tools: replace kdump-anaconda-addon with kexec-tools
I'm proposing making the addon part of anaconda-install-env-deps
instead in https://github.com/rhinstaller/anaconda/pull/5205 .
It's more correct. This group is meant to contain things that
may be needed *on the installed system*, not in the installer
environment. The addon is needed in the installer environment,
not the installed system.

However, one thing the addon does is add kexec-tools to the list
of packages to be installed to the installed system, so we should
add kexec-tools to this group for correctness.

Signed-off-by: Adam Williamson <awilliam@redhat.com>