F40: Anaconda-tools: Mark live media exclusions type="optional"
Anaconda-tools is intended to provide a set of packages that anaconda
may add to the list of installed packages to support some
functionality it is configuring on the machine. For example, on a grub
based system, it will add grubby, grub-efi, etc, while on a
systemd-boot system it will add sdubby, systemd-boot, etc.

For historical reasons, it has been installed on the live-media which
causes problems when there are conflicting packages in the group (ex
grub and systemd-boot). Beyond installing unnecessary packages (ex:
iscsi-initiator-utils on machines without iscsi) this also has the
unintended side effect that any changes for new features also install
those packages on existing machines in the field, even if they were
explicitly excluded from the live media.

Add a comment clarifying this behavior while also converting the
existing three excluded packages to type=optional.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>