F40: Add a base-graphical group and use it in Xorg-less KDE
base-graphical provides the lowest common denominator set of packages
for a desktop environment. At the moment this includes mesa drivers
and a graphical plymouth theme for the boot animation etc.

If your desktop is transitioning away from Xorg (base-x group)
or was concieved with Wayland, this is the group for you.