Add bash-color-prompt and anaconda-webui to critpath
anaconda's webui has been split out into a separate source
package, so we need to list it separately in the critical path.

I also suggest bash-color-prompt should be treated as critical
path. This will ensure it gets gated by openQA (so we can adjust
openQA to handle changes to the prompt before they break almost
every test for every update), but also it just makes sense: the
package is setting the default prompt, which could cause all
kinds of problems if there is a bad update to it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>