Add realmd to critical-path-server
It's part of domain enrolment (both FreeIPA and AD), which are
release-blocking features for Server. Today a realmd update
broke FreeIPA enrolment, which is causing tests for all Rawhide
critical path updates to fail; if realmd had been in critical
path itself we would have caught this and gated the update.

Signed-off-by: Adam Williamson <awilliam@redhat.com>