cleanup 2024-03: unreplaced retirements
Also removes the mongodb group entirely, because it isn't really
serving a purpose any more. It was originally meant to contain
the mongodb server itself, but that got removed years ago for
non-free licensing. There's no real use to a group which just
has a random client library in it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>