Run check-missing on development releases as part of CI
This is...wrong, but it's *usefully* wrong.

comps PRs aren't the only thing that can cause check-missing to
fail. They're not even the most *common* reason for check-missing
to fail. It's more likely to be caused by a package being retired
or a source package changing the binary packages it produces.

However, we cannot currently gate package retirements and/or
binary package set changes on this script. Doing so would require
a whole ton more work than just...doing this.

Doing this will cause us to notice problems the next time a comps
PR is touched. This is a definite improvement on us noticing
whenever I remember to run the script manually.

It will be somewhat annoying/inconvenient for people submitting
comps requests, but I think the benefit of problems in comps
getting noticed and fixed much more rapidly outweighs that.

Signed-off-by: Adam Williamson <awilliam@redhat.com>